requirements.txt
django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-rest-auth==5.0.1
django-allauth==0.57.0
# django-rest-allauth==1.1
# requests==2.31.0
urllib3==1.26.6